requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
Flask==2.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==9.0.1
openpyxl==3.1.2
pillow==11.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyperclip==1.9.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==1.0.1
pytweening==1.2.0
pywhatkit==5.4
requests==2.32.4
rubicon-objc==0.5.1
soupsieve==2.7
urllib3==2.5.0
Werkzeug==3.1.3
wikipedia==1.4.0
xlrd==2.0.2
